+++ bin/build_all.sh
@@ -32,5 +32,5 @@ CONFIGS=( cmip5.yaml cmip6.yaml access-om2.yaml access-cm2.yaml access-esm1-5.ya
 
 config_paths=( "${CONFIGS[@]/#/${CONFIG_DIR}/}" )
 
-metacat-build --build_base_path=${OUTPUT_BASE_PATH} --version=${version} ${config_paths[@]}
+catalog-build --build_base_path=${OUTPUT_BASE_PATH} --version=${version} ${config_paths[@]}
